New upstream version of Plupload. Small fix in Disqus module.
System version 0.361
modules:
Blogs 0.064
Comments 0.020
Cron 0.008
Disqus 1.0.1+build-2
Feedback 0.006
HybridAuth 2.1.1-cs29
OAuth2 0.027
Photo_gallery 0.015
Plupload 2.0.0-cs1
Static_pages 0.036
plugins:
Admin_default_theme 0.005
Fotorama 4.4.6-cs1
SimpleImage 2.3-cs1
TinyMCE 4.0.9-cs2

+++ components/modules/Blogs/Blogs.php
@@ -114,7 +114,7 @@ class Blogs extends Accessor {
 		/**
 		 * @var \cs\modules\Comments\Comments $Comments
 		 */
-		$data['comments_count']	= Config::instance()->module('Blogs')->enable_comments && $Comments ? $Comments->count($data['id']) : 0;
+		$data['comments_count']	= (int)(Config::instance()->module('Blogs')->enable_comments && $Comments ? $Comments->count($data['id']) : 0);
 		return $data;
 	}
 	/**
